move creator funcs into object, update tests, update imports
