+++ src/models/datePickerFactory/DatePickerFactoryTypes.ts
@@ -1,0 +1,7 @@
+import AtomicDateObject from "../AtomicDateObject";
+
+type AtomicDateObjectCreator = (index: number) => AtomicDateObject;
+
+export {
+  AtomicDateObjectCreator
+}
