+++ src/views/DatePickerListView.ts
@@ -2,7 +2,7 @@
  *  if continuous scroll is true or number of dates is more than 12 init virtual dom.
  *  if there is a virtual dom, then normal scrolling is disabled.
  */
-import DatePickerFactory from "../models/DatePickerFactory";
+import DatePickerFactory from "../models/datePickerFactory/DatePickerFactory";
 import DatePickerBaseView from "./DatePickerBaseView";
 import DatePickerView from "./DatePickerViewInterface";
 import AtomicDateObject from "../models/AtomicDateObject";
